Don't require torch ot import lance
