+++ python/python/benchmarks/test_read_write.py
@@ -10,10 +10,6 @@ import pyarrow.dataset as ds
 import pyarrow.parquet as pq
 from lance.file import LanceFileReader, LanceFileWriter
 
-# from lance.tracing import trace_to_chrome
-
-# trace_to_chrome(level="trace")
-
 """
 Data generation (in-memory)
 import duckdb
